requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
pytest
pytest-django
factory_boy
factory_boy
